refactor: inquiry dto 패키지 분리
